# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KPRI Profile API",
        "description": "API Documentation for KPRI Profile Application",
        "contact": {
            "name": "KPRI Admin",
            "email": "admin@kpri.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/faqs": {
            "get": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Get list of all FAQs",
                "description": "Returns list of all frequently asked questions",
                "operationId": "getFAQs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_faq": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "judul": {
                                                        "type": "string",
                                                        "example": "How do I reset my password?"
                                                    },
                                                    "deskripsi": {
                                                        "type": "string",
                                                        "example": "You can reset your password by clicking on the 'Forgot Password' link on the login page."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No FAQs found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No FAQs found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/faqs/{id}": {
            "get": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Get specific FAQ",
                "description": "Returns a specific FAQ by its ID",
                "operationId": "getFAQ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of FAQ to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id_faq": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "judul": {
                                                    "type": "string",
                                                    "example": "How do I reset my password?"
                                                },
                                                "deskripsi": {
                                                    "type": "string",
                                                    "example": "You can reset your password by clicking on the 'Forgot Password' link on the login page."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "FAQ not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "FAQ not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}